replace shouldStopBefore with stateless function
`shouldStopBefore()` was hard to change because it needed to be called
every time before extending a file's range by adding a point key. But in
the world of range data (i.e., range tombstones) this no longer makes
sense because we do not want a simple yes/no answer to whether a point
key is acceptable to add; instead, we want the exact key at which the
file should be cut as it may be in the middle of a range.

To provide this capability, this PR replaces `shouldStopBefore()` with a
function, `findGrandparentLimit()`. It takes the start key of a file and
returns the latest possible end key that can be used without violating
grandparent overlap rules. Unlike `shouldStopBefore()`, it does not
maintain any state about keys seen or files already passed, which I
believe makes it easier to maintain.

Also changed `finishOutput()` to take user key rather than internal key
as that's all it needs (and all `findGrandparentLimit()` returns).